spec/requirements.txt
concurrent-log-handler==0.9.20
Django==4.0.1
SQLAlchemy==1.4.37
coverage==6.4.1
django-filter==21.1
djangorestframework==3.13.1
mssql-django==1.1.3
pandas==1.4.2
django-cors-headers==3.11.0
django-auth-ldap==4.0.0
ldap3==2.9.1
